#33 renamed api routes
