+++ src/constants/routes.ts
@@ -2,8 +2,8 @@ export const apiRoutes = {
   "create-user": "/api/create-user",
   "confirm-email": "/api/confirm-email",
   "login-user": "/api/login-user",
-  "check-email": "/api/check-email",
-  "check-username": "/api/check-username",
+  "get-check-email": "/api/get-check-email",
+  "get-check-username": "/api/get-check-username",
   "forgot-password": "/api/forgot-password",
   "reset-password": "/api/reset-password",
   "get-current-weather": "/api/get-current-weather",
